feat: sequence 타입 변경 - Integer -> Long

+++ part5/pass-batch/src/main/java/com/fastcampus/pass/repository/booking/BookingRepository.java
@@ -14,5 +14,5 @@ public interface BookingRepository extends JpaRepository<BookingEntity, Long> {
             , b.modifiedAt = CURRENT_TIMESTAMP
             WHERE b.passSeq = :passSeq
             """)
-    int updateUsedPass(Integer passSeq, boolean usedPass);
+    int updateUsedPass(Long passSeq, boolean usedPass);
 }
